- Implemented support for MS identity abstractions (securitytoken, jwtsecuritytoken, etc) - Implement extension methods for easy conversion of security tokens to/from base64 - Updated flows to match new interface suggestion (review with Emil) - Streamlined internal class usage and proposed a solution that hides away some of the internals (partial classes) so its easier to work with the individual flows.
