--- src/ResourceProvisioning.Cli.Application/Authentication/AuthenticationToken.cs
@@ -1,11 +1,0 @@
-namespace ResourceProvisioning.Cli.Application.Authentication
-{
-	public class AuthenticationToken
-	{
-		public string IdToken { get; set; }
-		public string AccessToken { get; set; }
-		public string TokenType { get; set; }
-		public string Scope { get; set; }
-		public int ExpiresIn { get; set; }
-	}
-}
