+++ src/ResourceProvisioning.Cli.Application/Authentication/DeviceCodeFlow.cs
@@ -1,33 +1,36 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Net.Http;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
 
 namespace ResourceProvisioning.Cli.Application.Authentication
 {
-	public class DeviceCodeFlow : AuthenticationProvider
+	public partial class DeviceCodeFlow : AuthenticationProvider
 	{
 		public DeviceCodeFlow(IOptions<CliApplicationOptions> cliApplicationOptions) : base(cliApplicationOptions)
 		{
 		}
 
-		public override async Task<AuthenticationToken> Auth(params dynamic[] args)
+		public override async ValueTask<SecurityToken> Auth(SecurityTokenDescriptor descriptor = default)
 		{
 			var response = await InitialiseFlow();
 			var tokenResponse = await Poll(response);
 			
-			return new AuthenticationToken
-			{
-				AccessToken = tokenResponse.AccessToken,
-				IdToken = tokenResponse.IdToken,
-				Scope = tokenResponse.Scope,
-				ExpiresIn = tokenResponse.ExpiresIn,
-				TokenType = tokenResponse.TokenType
-			};
+			//TODO: Talk about this with Emil.
+			return new JwtSecurityToken(tokenResponse.IdToken);
+			//return new AuthenticationToken
+			//{
+			//	AccessToken = tokenResponse.AccessToken,
+			//	IdToken = tokenResponse.IdToken,
+			//	Scope = tokenResponse.Scope,
+			//	ExpiresIn = tokenResponse.ExpiresIn,
+			//	TokenType = tokenResponse.TokenType
+			//};
 		}
 		
 		private async Task<DeviceCodeResponse> InitialiseFlow()
@@ -43,8 +46,7 @@ namespace ResourceProvisioning.Cli.Application.Authentication
 			var deviceCodeResponse = JsonSerializer.Deserialize<DeviceCodeResponse>(respPayload);
 			return deviceCodeResponse;
 		}
-
-		
+				
 		private async Task<TokenValidResponse> Poll(DeviceCodeResponse deviceCodeResponse)
 		{
 			var dict = new Dictionary<string, string>();
@@ -84,55 +86,5 @@ namespace ResourceProvisioning.Cli.Application.Authentication
 				Thread.Sleep(deviceCodeResponse.Interval * 1000); // Wait the requested interval before polling again.
 			}
 		}
-	}
-	
-	class DeviceCodeResponse
-	{
-		[JsonPropertyName("user_code")]
-		public string UserCode { get; set; }
-		[JsonPropertyName("device_code")]
-		public string DeviceCode { get; set; }
-		[JsonPropertyName("verification_uri")]
-		public string VerificationUri { get; set; }
-		[JsonPropertyName("expires_in")]
-		public int ExpiresIn { get; set; }
-		[JsonPropertyName("interval")]
-		public int Interval { get; set; }
-		[JsonPropertyName("message")]
-		public string Message { get; set; }
-	}
-
-	class TokenValidResponse // 200
-	{
-		[JsonPropertyName("token_type")]
-		public string TokenType { get; set; }
-		[JsonPropertyName("scope")]
-		public string Scope { get; set; }
-		[JsonPropertyName("access_token")]
-		public string AccessToken { get; set; }
-		[JsonPropertyName("id_token")]
-		public string IdToken { get; set; }
-		[JsonPropertyName("expires_in")]
-		public int ExpiresIn { get; set; }
-		[JsonPropertyName("ext_expires_in")]
-		public int ExtExpiresIn { get; set; }
-	}
-
-	class TokenErrorResponse // 400
-	{
-		[JsonPropertyName("error")]
-		public string Error { get; set; }
-		[JsonPropertyName("error_description")]
-		public string ErrorDescription { get; set; }
-		[JsonPropertyName("error_codes")]
-		public IEnumerable<int> ErrorCodes { get; set; }
-		[JsonPropertyName("timestamp")]
-		public string Timestamp { get; set; }
-		[JsonPropertyName("trace_id")]
-		public string TraceId { get; set; }
-		[JsonPropertyName("correlation_id")]
-		public string CorrelationId { get; set; }
-		[JsonPropertyName("error_uri")]
-		public string ErrorUri { get; set; }
 	}
 }
