+++ src/ResourceProvisioning.Cli.Application/Authentication/PassThruFlow.cs
@@ -1,20 +1,25 @@
-﻿using System.Net.Http.Headers;
+﻿using System.IdentityModel.Tokens.Jwt;
+using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
 
 namespace ResourceProvisioning.Cli.Application.Authentication
 {
-	public class PassThruFlow : AuthenticationProvider<AuthenticationToken>
+	public class PassThruFlow : AuthenticationProvider
 	{
 		public PassThruFlow(IOptions<CliApplicationOptions> cliApplicationOptions) : base(cliApplicationOptions)
 		{
 		}
 
-		public override ValueTask<AuthenticationToken> Auth(AuthenticationToken args = default)
+		public override ValueTask<SecurityToken> Auth(SecurityTokenDescriptor descriptor = default)
 		{
-			HttpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("bearer", args?.AccessToken);
+			//TODO: Fix these values later.
+			var jwt = new JwtSecurityToken("ssucli", "audience").Convert(new JwtSecurityTokenHandler(), descriptor);
 
-            return new ValueTask<AuthenticationToken>(args);
+			HttpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(jwt.GetAuthenticationScheme(), jwt.ToBase64String());
+
+            return new ValueTask<SecurityToken>(jwt);
 		}
 	}
 }
