--- src/ResourceProvisioning.Cli.Application/Authentication/UserData.cs
@@ -1,8 +1,0 @@
-namespace ResourceProvisioning.Cli.Application.Authentication
-{
-	class UserData
-	{
-		public string Username { get; set; }
-		public string Password { get; set; }
-	}
-}
