+++ src/ResourceProvisioning.Cli.Application/ResourceProvisioning.Cli.Application.csproj
@@ -15,6 +15,7 @@
       <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="3.1.5" />
       <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.5" />
       <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="3.1.5" />
+      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.7.1" />
     </ItemGroup>
 
     <ItemGroup>
